requirements.txt
# Data Processing
pandas
openpyxl
sqlalchemy
xlsxwriter
pyxlsb
tqdm

# Environment
python-dotenv

# Strands RCSA_Agents with OpenTelemetry and Anthropic support
strands-agents[otel,anthropic]
strands-agents-tools
strands-agents-builder

# Jupyter/Notebooks
ipywidgets

# FastAPI Backend
fastapi
uvicorn[standard]
sse-starlette
python-multipart
pydantic>=2.0
pydantic-settings

# ML / Statistics
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.5
statsmodels>=0.14.0
scipy>=1.10.0
optbinning>=0.17.0

# Report Generation
python-docx>=0.8.11
matplotlib>=3.6.0
seaborn>=0.12.0

# Model Persistence
joblib>=1.2.0

# Explainability
shap>=0.42.0

# LLM Providers
anthropic
openai
openai-agents
tiktoken
claude-agent-sdk
claude-code-sdk